Move math tests to seperate test classes
