+++ src/test/java/dk/alexandra/fresco/lib/arithmetic/BasicArithmeticTests.java
@@ -28,7 +28,6 @@ package dk.alexandra.fresco.lib.arithmetic;
 
 import java.io.IOException;
 import java.math.BigInteger;
-import java.util.Arrays;
 
 import org.junit.Assert;
 
@@ -42,33 +41,11 @@ import dk.alexandra.fresco.framework.sce.SCE;
 import dk.alexandra.fresco.framework.sce.SCEFactory;
 import dk.alexandra.fresco.framework.value.OInt;
 import dk.alexandra.fresco.framework.value.SInt;
-import dk.alexandra.fresco.lib.compare.MiscOIntGenerators;
-import dk.alexandra.fresco.lib.compare.RandomAdditiveMaskFactory;
-import dk.alexandra.fresco.lib.compare.RandomAdditiveMaskFactoryImpl;
 import dk.alexandra.fresco.lib.field.integer.BasicNumericFactory;
 import dk.alexandra.fresco.lib.helper.CopyProtocolImpl;
 import dk.alexandra.fresco.lib.helper.builder.NumericIOBuilder;
 import dk.alexandra.fresco.lib.helper.builder.NumericProtocolBuilder;
 import dk.alexandra.fresco.lib.helper.sequential.SequentialProtocolProducer;
-import dk.alexandra.fresco.lib.math.integer.PreprocessedNumericBitFactory;
-import dk.alexandra.fresco.lib.math.integer.binary.RepeatedRightShiftProtocol;
-import dk.alexandra.fresco.lib.math.integer.binary.RightShiftFactory;
-import dk.alexandra.fresco.lib.math.integer.binary.RightShiftFactoryImpl;
-import dk.alexandra.fresco.lib.math.integer.binary.RightShiftProtocol;
-import dk.alexandra.fresco.lib.math.integer.division.DivisionFactory;
-import dk.alexandra.fresco.lib.math.integer.division.DivisionFactoryImpl;
-import dk.alexandra.fresco.lib.math.integer.division.DivisionProtocol;
-import dk.alexandra.fresco.lib.math.integer.linalg.InnerProductFactory;
-import dk.alexandra.fresco.lib.math.integer.linalg.InnerProductFactoryImpl;
-import dk.alexandra.fresco.lib.math.integer.sqrt.SquareRootFactory;
-import dk.alexandra.fresco.lib.math.integer.sqrt.SquareRootFactoryImpl;
-import dk.alexandra.fresco.lib.math.integer.sqrt.SquareRootProtocol;
-import dk.alexandra.fresco.lib.math.integer.stat.CovarianceMatrixProtocol;
-import dk.alexandra.fresco.lib.math.integer.stat.CovarianceProtocol;
-import dk.alexandra.fresco.lib.math.integer.stat.MeanProtocol;
-import dk.alexandra.fresco.lib.math.integer.stat.StatisticsFactory;
-import dk.alexandra.fresco.lib.math.integer.stat.StatisticsFactoryImpl;
-import dk.alexandra.fresco.lib.math.integer.stat.VarianceProtocol;
 
 
 /**
@@ -573,451 +550,6 @@ public class BasicArithmeticTests {
 						}
 					};
 					sce.runApplication(app);					
-				}
-			};
-		}
-	}
-	
-	/**
-	 * Test binary right shift of a shared secret.
-	 */
-	public static class TestRightShift extends TestThreadFactory {
-
-		@Override
-		public TestThread next(TestThreadConfiguration conf) {
-			
-			return new ThreadWithFixture() {
-				private final BigInteger input = BigInteger.valueOf(12332157);
-
-				@Override
-				public void test() throws Exception {
-					TestApplication app = new TestApplication() {
-
-						private static final long serialVersionUID = 701623441111137585L;
-						
-						@Override
-						public ProtocolProducer prepareApplication(
-								ProtocolFactory provider) {
-							
-							BasicNumericFactory basicNumericFactory = (BasicNumericFactory) provider;
-							PreprocessedNumericBitFactory preprocessedNumericBitFactory = (PreprocessedNumericBitFactory) provider;
-							RandomAdditiveMaskFactory randomAdditiveMaskFactory = new RandomAdditiveMaskFactoryImpl(basicNumericFactory, preprocessedNumericBitFactory);
-							RightShiftFactory rightShiftFactory = new RightShiftFactoryImpl(80, basicNumericFactory, randomAdditiveMaskFactory);
-
-							SInt result = basicNumericFactory.getSInt();
-							SInt remainder = basicNumericFactory.getSInt();
-
-							NumericIOBuilder ioBuilder = new NumericIOBuilder(basicNumericFactory);
-							SequentialProtocolProducer sequentialProtocolProducer = new SequentialProtocolProducer();
-							
-							SInt input1 = ioBuilder.input(input, 1);
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							RightShiftProtocol rightShiftProtocol = rightShiftFactory.getRightShiftProtocol(input1, result, remainder);
-							sequentialProtocolProducer.append(rightShiftProtocol);
-							
-							OInt output1 = ioBuilder.output(result);
-							OInt output2 = ioBuilder.output(remainder);
-							
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							ProtocolProducer gp = sequentialProtocolProducer;
-							
-							outputs = new OInt[] {output1, output2};
-							
-							return gp;
-						}
-					};
-					sce.runApplication(app);
-					BigInteger result = app.getOutputs()[0].getValue();
-					BigInteger remainder = app.getOutputs()[1].getValue();
-					System.out.println(input + " >> 1 = " + result);
-					System.out.println(input + " (mod 2) = " + remainder);
-					
-					Assert.assertEquals(result, input.shiftRight(1));
-					Assert.assertEquals(remainder, input.mod(BigInteger.valueOf(2)));
-				}
-			};
-		}
-	}
-	
-
-	public static class TestRepeatedRightShift extends TestThreadFactory {
-
-		@Override
-		public TestThread next(TestThreadConfiguration conf) {
-			
-			return new ThreadWithFixture() {
-				private final BigInteger input = BigInteger.valueOf(12332153);
-				private final int n = 7;
-				
-				@Override
-				public void test() throws Exception {
-					TestApplication app = new TestApplication() {
-
-						private static final long serialVersionUID = 701623441111137585L;
-						
-						@Override
-						public ProtocolProducer prepareApplication(
-								ProtocolFactory provider) {
-							
-							BasicNumericFactory basicNumericFactory = (BasicNumericFactory) provider;
-							PreprocessedNumericBitFactory preprocessedNumericBitFactory = (PreprocessedNumericBitFactory) provider;
-							RandomAdditiveMaskFactory randomAdditiveMaskFactory = new RandomAdditiveMaskFactoryImpl(basicNumericFactory, preprocessedNumericBitFactory);
-							RightShiftFactory rightShiftFactory = new RightShiftFactoryImpl(80, basicNumericFactory, randomAdditiveMaskFactory);
-
-							SInt result = basicNumericFactory.getSInt();
-							
-							SInt[] remainders = new SInt[n];
-							for (int i = 0; i < n; i++) {
-								remainders[i] = basicNumericFactory.getSInt();
-							}
-							
-							NumericIOBuilder ioBuilder = new NumericIOBuilder(basicNumericFactory);
-							SequentialProtocolProducer sequentialProtocolProducer = new SequentialProtocolProducer();
-							
-							SInt input1 = ioBuilder.input(input, 1);
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							RepeatedRightShiftProtocol rightShiftProtocol = rightShiftFactory.getRepeatedRightShiftProtocol(input1, n, result, remainders);
-							sequentialProtocolProducer.append(rightShiftProtocol);
-							
-							OInt shiftOutput = ioBuilder.output(result);
-							OInt[] remainderOutput = ioBuilder.outputArray(remainders);
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							ProtocolProducer gp = sequentialProtocolProducer;
-							
-							outputs = new OInt[n+1];
-							outputs[0] = shiftOutput;
-							for (int i = 0; i < n; i++) { 
-								outputs[i+1] = remainderOutput[i];
-							}
-							return gp;
-						}
-					};
-					sce.runApplication(app);
-					
-					BigInteger output = app.getOutputs()[0].getValue();
-					System.out.println(input + " >> " + n + " = " + output);
-					Assert.assertEquals(input.shiftRight(n), output);
-
-					BigInteger[] remainders = new BigInteger[n];
-					for (int i = 0; i < n; i++) {
-						remainders[i] = app.getOutputs()[i+1].getValue();
-						Assert.assertEquals(input.testBit(i) ? BigInteger.ONE : BigInteger.ZERO, remainders[i]);
-					}
-					System.out.println(Arrays.toString(remainders));
-				}
-			};
-		}
-	}
-
-	/**
-	 * Test Euclidian division
-	 */
-	public static class TestEuclidianDivision extends TestThreadFactory {
-
-		@Override
-		public TestThread next(TestThreadConfiguration conf) {
-			
-			return new ThreadWithFixture() {
-				private final BigInteger x = new BigInteger("123978634193227335452345761");
-				private final BigInteger d = new BigInteger("6543212341214412");
-
-				@Override
-				public void test() throws Exception {
-					TestApplication app = new TestApplication() {
-
-						private static final long serialVersionUID = 701623441111137585L;
-						
-						@Override
-						public ProtocolProducer prepareApplication(
-								ProtocolFactory provider) {
-							
-							BasicNumericFactory basicNumericFactory = (BasicNumericFactory) provider;
-							PreprocessedNumericBitFactory preprocessedNumericBitFactory = (PreprocessedNumericBitFactory) provider;
-							RandomAdditiveMaskFactory randomAdditiveMaskFactory = new RandomAdditiveMaskFactoryImpl(basicNumericFactory, preprocessedNumericBitFactory);
-							MiscOIntGenerators miscOIntGenerators = new MiscOIntGenerators(basicNumericFactory);
-							InnerProductFactory innerProductFactory = new InnerProductFactoryImpl(basicNumericFactory);
-							RightShiftFactory rightShiftFactory = new RightShiftFactoryImpl(80, basicNumericFactory, randomAdditiveMaskFactory);
-							DivisionFactory euclidianDivisionFactory = new DivisionFactoryImpl(basicNumericFactory, rightShiftFactory);
-							
-							SInt quotient = basicNumericFactory.getSInt();
-							SInt remainder = basicNumericFactory.getSInt();
-
-							NumericIOBuilder ioBuilder = new NumericIOBuilder(basicNumericFactory);
-							SequentialProtocolProducer sequentialProtocolProducer = new SequentialProtocolProducer();
-							
-							SInt input1 = ioBuilder.input(x, 1);
-							OInt input2 = basicNumericFactory.getOInt(d);
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							DivisionProtocol euclidianDivisionProtocol = euclidianDivisionFactory.getDivisionProtocol(input1, x.bitLength() + 1, input2, quotient, remainder);
-							sequentialProtocolProducer.append(euclidianDivisionProtocol);
-							
-							OInt output1 = ioBuilder.output(quotient);
-							OInt output2 = ioBuilder.output(remainder);
-							
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							ProtocolProducer gp = sequentialProtocolProducer;
-							
-							outputs = new OInt[] {output1, output2};
-							
-							return gp;
-						}
-					};
-					sce.runApplication(app);
-					BigInteger quotient = app.getOutputs()[0].getValue();
-					BigInteger remainder = app.getOutputs()[1].getValue();
-					System.out.println(x + " / " + d + " = " + quotient + " with remainder " + remainder);
-					Assert.assertEquals(quotient, x.divide(d));
-					Assert.assertEquals(remainder, x.mod(d));
-				}
-			};
-		}
-	}
-
-	/**
-	 * Test Euclidian division
-	 */
-	public static class TestStatistics extends TestThreadFactory {
-
-		@Override
-		public TestThread next(TestThreadConfiguration conf) {
-			
-			return new ThreadWithFixture() {
-				private final int[] data1 = {543,520,532,497,450,432};
-				private final int[] data2 = {432,620,232,337,250,433};
-				private final int[] data3 = {80,90,123,432,145,606};
-				int count = 3;
-				
-				@Override
-				public void test() throws Exception {
-					TestApplication app = new TestApplication() {
-
-						private static final long serialVersionUID = 701623441111137585L;
-						
-						@Override
-						public ProtocolProducer prepareApplication(
-								ProtocolFactory provider) {
-							
-							BasicNumericFactory basicNumericFactory = (BasicNumericFactory) provider;
-							PreprocessedNumericBitFactory preprocessedNumericBitFactory = (PreprocessedNumericBitFactory) provider;
-							RandomAdditiveMaskFactory randomAdditiveMaskFactory = new RandomAdditiveMaskFactoryImpl(basicNumericFactory, preprocessedNumericBitFactory);
-							RightShiftFactory rightShiftFactory = new RightShiftFactoryImpl(80, basicNumericFactory, randomAdditiveMaskFactory);
-							DivisionFactory euclidianDivisionFactory = new DivisionFactoryImpl(basicNumericFactory, rightShiftFactory);
-							StatisticsFactory statisticsFactory = new StatisticsFactoryImpl(basicNumericFactory, euclidianDivisionFactory);
-							
-							NumericIOBuilder ioBuilder = new NumericIOBuilder(basicNumericFactory);
-							SequentialProtocolProducer sequentialProtocolProducer = new SequentialProtocolProducer();
-							
-							SInt mean1 = basicNumericFactory.getSInt();
-							SInt mean2 = basicNumericFactory.getSInt();
-							SInt variance = basicNumericFactory.getSInt();
-							SInt covariance = basicNumericFactory.getSInt();
-							
-							SInt[][] covarianceMatrix = new SInt[3][3];
-							for (int i = 0; i < 3; i++) {
-								for (int j = 0; j < 3; j++) {
-									covarianceMatrix[i][j] = basicNumericFactory.getSInt();
-								}
-							}
-							
-							SInt[][] input = ioBuilder.inputMatrix(new int[][] {data1, data2, data3}, 1);
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							MeanProtocol arithmeticMeanProtocol = statisticsFactory.getMeanProtocol(input[0], 10, mean1);
-							sequentialProtocolProducer.append(arithmeticMeanProtocol);
-
-							MeanProtocol arithmeticMeanProtocol2 = statisticsFactory.getMeanProtocol(input[1], 10, mean2);
-							sequentialProtocolProducer.append(arithmeticMeanProtocol2);
-
-							VarianceProtocol varianceProtocol = statisticsFactory.getVarianceProtocol(input[0], 10, mean1, variance);
-							sequentialProtocolProducer.append(varianceProtocol);
-							
-							CovarianceProtocol covarianceProtocol = statisticsFactory.getCovarianceProtocol(input[0], input[1], 10, mean1, mean2, covariance);
-							sequentialProtocolProducer.append(covarianceProtocol);
-							
-							CovarianceMatrixProtocol covarianceMatrixProtocol = statisticsFactory.getCovarianceMatrixProtocol(input, 10, covarianceMatrix);
-							sequentialProtocolProducer.append(covarianceMatrixProtocol);
-							
-							OInt output1 = ioBuilder.output(mean1);
-							OInt output2 = ioBuilder.output(mean2);
-							OInt output3 = ioBuilder.output(variance);
-							OInt output4 = ioBuilder.output(covariance);
-							OInt[][] output5 = ioBuilder.outputMatrix(covarianceMatrix);
-							
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							ProtocolProducer gp = sequentialProtocolProducer;
-							
-							outputs = new OInt[] {output1, output2, output3, output4, output5[0][0], output5[1][0]};
-							
-							return gp;
-						}
-					};
-					sce.runApplication(app);
-					BigInteger mean1 = app.getOutputs()[0].getValue();
-					BigInteger mean2 = app.getOutputs()[1].getValue();
-					BigInteger variance = app.getOutputs()[2].getValue();
-					BigInteger covariance = app.getOutputs()[3].getValue();
-					
-					double sum = 0.0;
-					for (int entry : data1) {
-						sum += entry;
-					}
-					double mean1Exact = sum / data1.length;
-							
-					sum = 0.0;
-					for (int entry : data2) {
-						sum += entry;
-					}
-					double mean2Exact = sum / data2.length;
-					
-					
-					double ssd = 0.0;
-					for (int entry : data1) {
-						ssd += (entry - mean1Exact) * (entry - mean1Exact);
-					}
-					double varianceExact = ssd / (data1.length - 1);
-					
-					double covarianceExact = 0.0;
-					for (int i = 0; i < data1.length; i++) {
-						covarianceExact += (data1[i] - mean1Exact) * (data2[i] - mean2Exact);
-					}
-					covarianceExact /= (data1.length - 1);
-					
-					System.out.println("Calculated mean1 = " + mean1 + " (exact value is " + mean1Exact + ")");
-					System.out.println("Calculated mean2 = " + mean2 + " (exact value is " + mean2Exact + ")");
-					System.out.println("Calculated variance = " + variance + " (exact value is " + varianceExact + ")");
-					System.out.println("Calculated covariance = " + covariance + " (exact value is " + covarianceExact + ")");
-					
-					System.out.println("Cov[0][0] = " + app.getOutputs()[4].getValue() + " (should be " + varianceExact + ")");
-					System.out.println("Cov[0][1] = " + app.getOutputs()[5].getValue() + " (should be " + covarianceExact + ")");
-
-				}
-			};
-		}
-	}
-	
-	/**
-	 * Test Euclidian division
-	 */
-	public static class TestSecretSharedDivision extends TestThreadFactory {
-
-		@Override
-		public TestThread next(TestThreadConfiguration conf) {
-			
-			return new ThreadWithFixture() {
-				private final BigInteger x = new BigInteger("8000");
-				private final BigInteger d = new BigInteger("3");
-				private final int precision = 4; // How many bits of precision do we get? Should be 2^p / l...
-
-				@Override
-				public void test() throws Exception {
-					TestApplication app = new TestApplication() {
-
-						private static final long serialVersionUID = 701623441111137585L;
-						
-						@Override
-						public ProtocolProducer prepareApplication(
-								ProtocolFactory provider) {
-							
-							BasicNumericFactory basicNumericFactory = (BasicNumericFactory) provider;
-							PreprocessedNumericBitFactory preprocessedNumericBitFactory = (PreprocessedNumericBitFactory) provider;
-							RandomAdditiveMaskFactory randomAdditiveMaskFactory = new RandomAdditiveMaskFactoryImpl(basicNumericFactory, preprocessedNumericBitFactory);
-							MiscOIntGenerators miscOIntGenerators = new MiscOIntGenerators(basicNumericFactory);
-							InnerProductFactory innerProductFactory = new InnerProductFactoryImpl(basicNumericFactory);
-							RightShiftFactory rightShiftFactory = new RightShiftFactoryImpl(80, basicNumericFactory, randomAdditiveMaskFactory);
-							DivisionFactory divisionFactory = new DivisionFactoryImpl(basicNumericFactory, rightShiftFactory);
-							
-							SInt quotient = basicNumericFactory.getSInt();
-
-							NumericIOBuilder ioBuilder = new NumericIOBuilder(basicNumericFactory);
-							SequentialProtocolProducer sequentialProtocolProducer = new SequentialProtocolProducer();
-							
-							SInt input1 = ioBuilder.input(x, 1);
-							SInt input2 = ioBuilder.input(d, 2);
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							DivisionProtocol divisionProtocol = divisionFactory.getDivisionProtocol(input1, input2, d.bitLength() + 1, precision, quotient);
-							sequentialProtocolProducer.append(divisionProtocol);
-							
-							OInt output1 = ioBuilder.output(quotient);
-							
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							ProtocolProducer gp = sequentialProtocolProducer;
-							
-							outputs = new OInt[] {output1};
-							
-							return gp;
-						}
-					};
-					sce.runApplication(app);
-					BigInteger quotient = app.getOutputs()[0].getValue();
-					System.out.println(x + " / " + d + " = " + quotient);
-					Assert.assertEquals(quotient, x.divide(d));
-				}
-			};
-		}
-	}
-
-	public static class TestSquareRoot extends TestThreadFactory {
-
-		@Override
-		public TestThread next(TestThreadConfiguration conf) {
-			
-			return new ThreadWithFixture() {
-				private final BigInteger x = new BigInteger("423402342001");
-
-				@Override
-				public void test() throws Exception {
-					TestApplication app = new TestApplication() {
-
-						private static final long serialVersionUID = 701623441111137585L;
-						
-						@Override
-						public ProtocolProducer prepareApplication(
-								ProtocolFactory provider) {
-							
-							BasicNumericFactory basicNumericFactory = (BasicNumericFactory) provider;
-							PreprocessedNumericBitFactory preprocessedNumericBitFactory = (PreprocessedNumericBitFactory) provider;
-							RandomAdditiveMaskFactory randomAdditiveMaskFactory = new RandomAdditiveMaskFactoryImpl(basicNumericFactory, preprocessedNumericBitFactory);
-							MiscOIntGenerators miscOIntGenerators = new MiscOIntGenerators(basicNumericFactory);
-							InnerProductFactory innerProductFactory = new InnerProductFactoryImpl(basicNumericFactory);
-							RightShiftFactory rightShiftFactory = new RightShiftFactoryImpl(80, basicNumericFactory, randomAdditiveMaskFactory);
-							DivisionFactory divisionFactory = new DivisionFactoryImpl(basicNumericFactory, rightShiftFactory);
-							SquareRootFactory squareRootFactory = new SquareRootFactoryImpl(basicNumericFactory, divisionFactory, rightShiftFactory);
-							
-							SInt sqrt = basicNumericFactory.getSInt();
-
-							NumericIOBuilder ioBuilder = new NumericIOBuilder(basicNumericFactory);
-							SequentialProtocolProducer sequentialProtocolProducer = new SequentialProtocolProducer();
-							
-							SInt input1 = ioBuilder.input(x, 1);
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							SquareRootProtocol squareRootProtocol = squareRootFactory.getSquareRootProtocol(input1, x.bitLength(), sqrt);
-							sequentialProtocolProducer.append(squareRootProtocol);
-							
-							OInt output1 = ioBuilder.output(sqrt);
-							
-							sequentialProtocolProducer.append(ioBuilder.getCircuit());
-							
-							ProtocolProducer gp = sequentialProtocolProducer;
-							
-							outputs = new OInt[] {output1};
-							
-							return gp;
-						}
-					};
-					sce.runApplication(app);
-					BigInteger sqrt = app.getOutputs()[0].getValue();
-					System.out.println("sqrt(" + x + ") = " + sqrt);
-					Assert.assertEquals(sqrt, BigInteger.valueOf(650693));
 				}
 			};
 		}
